requirements.txt
pyserial==3.0.1
requests==2.9.1
